[master][537539](UserStory) update generated client api code

+++ flixpy/flix/extension/extension_api/models/episode_details_dto.py
@@ -12,14 +12,14 @@ T = TypeVar("T", bound="EpisodeDetailsDto")
 class EpisodeDetailsDto:
     """
     Attributes:
-        id (float):
+        id (int):
         tracking_code (str):
         created_date (datetime.datetime):
         owner (str):
         title (Optional[str]):
     """
 
-    id: float
+    id: int
     tracking_code: str
     created_date: datetime.datetime
     owner: str
